Update create_share_links.sh
Small addition that will check links before proceeding to make links again. The process errors out without this if one just tries to rebuild again without deleting the existing resource group. I have run into this situation.

+++ examples/reveal_E2E/scripts/create_share_links.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
-
+if [ ! -L /share/apps ]
+then
 ln -s /apps /share/apps
 ln -s /data /share/data
+fi
